JS structure. Improved.
