+++ index.html
@@ -156,7 +156,7 @@
     </style>
   </head>
   <body>
-    <div id="loading" class="spinner" style="visibility: visible;"></div>
+    <div id="loading" class="spinner visible"></div>
     <button id="checkAvailabilityBtn">
       <span>Check Availability</span>
     </button>
