+++ src/styles.scss
@@ -1,7 +1,15 @@
-$primary-color: #861ea0;
+$primary-color: #cccad9;
 
 body {
   font-family: Arial, sans-serif;
   background-color: $primary-color;
-  color: white;
+  color: black;
+}
+
+.visible {
+  visibility: visible;
+}
+
+.hidden {
+  visibility: hidden;
 }
